Implement Slack-style messages view
Before this change, the first message in a conversation goes to
the top of the messages view. Subsequent messages get added beneath
existing messages. When it comes to scrolling you're on your own.

After this change, the first message in a conversation goes to the
bottom of the messages view. The stack of messages grows taller as
new messages are added to the bottom. Scrolling to the bottom "pins"
the window to the bottom of the list, so as to keep the user up to
date with new messages.